[BUGFIX] Refactor to restore 6.2 support
Technically, the extension does support 6.2 with older versions of cmis_service.

The FAL driver change happened in 7.6, not v8 - refactored accordingly.
