+++ ext_emconf.php
@@ -32,7 +32,7 @@ $EM_CONF[$_EXTKEY] = array(
     'version' => '1.0.0',
 	'constraints' => array(
 		'depends' => array(
-			'typo3' => '7.6.0-8.7.99'
+			'typo3' => '6.2.11-8.7.99'
 		),
 		'conflicts' => array(
 		),
