+++ ext_localconf.php
@@ -3,10 +3,10 @@ defined('TYPO3_MODE') or die('Access denied');
 
 (function() {
 
-    $driverClassName = \Dkd\CmisFal\Driver\Versioned\V7\CMISFilesystemDriver::class;
+    $driverClassName = \Dkd\CmisFal\Driver\Versioned\V6\CMISFilesystemDriver::class;
 
-    if (version_compare(\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::getExtensionVersion('core'), 8, '>=')) {
-        $driverClassName = \Dkd\CmisFal\Driver\Versioned\V8\CMISFilesystemDriver::class;
+    if (version_compare(\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::getExtensionVersion('core'), "7.6", '>=')) {
+        $driverClassName = \Dkd\CmisFal\Driver\Versioned\V7\CMISFilesystemDriver::class;
     }
 
     $GLOBALS['TYPO3_CONF_VARS']['SYS']['fal']['registeredDrivers']['cmis'] = array(
